# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema Inmobiliario",
        "description": "Documentación de la API para gestión de propiedades, inquilinos y contratos.",
        "contact": {
            "email": "admin@test.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/inquilinos": {
            "get": {
                "tags": [
                    "Inquilinos"
                ],
                "summary": "Listar todos los inquilinos",
                "operationId": "89366977c743a1ab807e5674d9b999f4",
                "responses": {
                    "200": {
                        "description": "Lista de inquilinos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inquilinos"
                ],
                "summary": "Registrar un nuevo inquilino",
                "operationId": "0feb896b9f949b81450b32e6088f079b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_completo",
                                    "dni"
                                ],
                                "properties": {
                                    "nombre_completo": {
                                        "type": "string",
                                        "example": "Carlos Pérez"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "30123456"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos@mail.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "3454123456"
                                    },
                                    "garante_nombre": {
                                        "type": "string",
                                        "example": "Padre de Carlos"
                                    },
                                    "garante_dni": {
                                        "type": "string",
                                        "example": "20123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inquilino creado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/inquilinos/{id}": {
            "put": {
                "tags": [
                    "Inquilinos"
                ],
                "summary": "Actualizar un inquilino existente",
                "operationId": "fd19cef6415abc6dd6001a035b90993e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del inquilino",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre_completo": {
                                        "type": "string"
                                    },
                                    "dni": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inquilino actualizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inquilinos"
                ],
                "summary": "Eliminar un inquilino",
                "operationId": "ad277b65fbd6b714ab24bfd356a6acf3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inquilino eliminado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Login con token Bearer",
                "name": "Token de Acceso",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}